outsourced calculation of relative humidity into utils
