+++ reskit/dac/workflows/dac_workflow_manager.py
@@ -56,37 +56,6 @@ class DACWorkflowManager(WorkflowManager):
         }
         self.units = OrderedDict(units)
 
-    def calculate_relative_humidity(self):
-        """
-        Function to calculate the relative humidity from dewpoint temperature and air temperature using the Sonntag formula.
-
-        References:
-        -------------
-        [1] https://www.npl.co.uk/resources/q-a/dew-point-and-relative-humidity
-        """
-
-        def calculate_vapor_pressure(temperature):
-            """
-            Function to calculate the vapor pressure from the temperature
-
-            temperature: temperature in °C
-            """
-            temperature_Kelvin = temperature + 273.15
-            vapor_pressure = np.exp(
-                -6096.9385 / temperature_Kelvin
-                + 21.2409642
-                - 2.711193 * 10**-2 * temperature_Kelvin
-                + 1.673952 * 10**-5 * temperature_Kelvin**2
-                + 2.433502 * np.log(temperature_Kelvin)
-            )  # vapor pressure [1]
-            return vapor_pressure
-
-        self.sim_data["relative_humidity"] = (
-            calculate_vapor_pressure(self.sim_data["surface_dew_temperature"])
-            / calculate_vapor_pressure(self.sim_data["surface_air_temperature"])
-            * 100
-        )  # [1]
-        del self.sim_data["surface_dew_temperature"]
 
     def load_lt_dac_model_data(self, model):
         """
